refactor(reporter): Code cleanup.
