+++ lib/reporter.js
@@ -1,38 +1,31 @@
+// Coverage Reporter
+// =====================
+//
+// Generates the report
+
+// Dependencies
+// ------------
+
 var path = require('path')
-var util = require('util')
 var istanbul = require('istanbul')
+
 var globalSourceCache = require('./sourceCache')
 var coverageMap = require('./coverageMap')
-
-var Store = istanbul.Store
-
-var SourceCacheStore = function (opts) {
-  Store.call(this, opts)
-  opts = opts || {}
-  this.sourceCache = opts.sourceCache
-}
-SourceCacheStore.TYPE = 'sourceCacheLookup'
-util.inherits(SourceCacheStore, Store)
-
-Store.mix(SourceCacheStore, {
-  keys: function () {
-    throw new Error('Not implemented')
-  },
-  get: function (key) {
-    return this.sourceCache[key]
-  },
-  hasKey: function (key) {
-    return this.sourceCache.hasOwnProperty(key)
-  },
-  set: function (key, contents) {
-    throw new Error('Not applicable')
-  }
-})
+var SourceCacheStore = require('./source-cache-store')
 
 // TODO(vojta): inject only what required (config.basePath, config.coverageReporter)
 var CoverageReporter = function (rootConfig, helper, logger) {
   var _ = helper._
   var log = logger.create('coverage')
+
+  // Instance variables
+  // ------------------
+
+  this.adapters = []
+
+  // Options
+  // -------
+
   var config = rootConfig.coverageReporter || {}
   var basePath = rootConfig.basePath
   var reporters = config.reporters
@@ -47,7 +40,6 @@ var CoverageReporter = function (rootConfig, helper, logger) {
     reporters = [config]
   }
 
-  this.adapters = []
   var collectors
   var pendingFileWritings = 0
   var fileWritingFinished = function () {}
@@ -64,28 +56,21 @@ var CoverageReporter = function (rootConfig, helper, logger) {
 
   function disposeCollectors () {
     if (pendingFileWritings <= 0) {
-      Object.keys(collectors).forEach(function (key) {
-        collectors[key].dispose()
+      _.forEach(collectors, function (collector) {
+        collector.dispose()
       })
+
       fileWritingFinished()
     }
   }
 
-  /**
-   * Generate the output directory from the `coverageReporter.dir` and
-   * `coverageReporter.subdir` options.
-   *
-   * @param {String} browserName - The browser name
-   * @param {String} dir - The given option
-   * @param {String|Function} subdir - The given option
-   *
-   * @return {String} - The output directory
-   */
+  // Generate the output directory from the `coverageReporter.dir` and
+  // `coverageReporter.subdir` options.
   function generateOutputDir (browserName, dir, subdir) {
     dir = dir || 'coverage'
     subdir = subdir || browserName
 
-    if (typeof subdir === 'function') {
+    if (_.isFunction(subdir)) {
       subdir = subdir(browserName)
     }
 
@@ -103,64 +88,68 @@ var CoverageReporter = function (rootConfig, helper, logger) {
 
   this.onBrowserStart = function (browser) {
     collectors[browser.id] = new istanbul.Collector()
-    if (includeAllSources) {
-      collectors[browser.id].add(coverageMap.get())
-    }
+
+    if (!includeAllSources) return
+
+    collectors[browser.id].add(coverageMap.get())
   }
 
   this.onBrowserComplete = function (browser, result) {
     var collector = collectors[browser.id]
 
-    if (!collector) {
-      return
-    }
+    if (!collector) return
+    if (!result || !result.coverage) return
 
-    if (result && result.coverage) {
-      collector.add(result.coverage)
-    }
+    collector.add(result.coverage)
   }
 
   this.onSpecComplete = function (browser, result) {
-    if (result.coverage) {
-      collectors[browser.id].add(result.coverage)
-    }
+    if (!result.coverage) return
+
+    collectors[browser.id].add(result.coverage)
   }
 
   this.onRunComplete = function (browsers) {
     reporters.forEach(function (reporterConfig) {
       browsers.forEach(function (browser) {
         var collector = collectors[browser.id]
-        if (collector) {
-          pendingFileWritings++
 
-          var mainDir = reporterConfig.dir || config.dir
-          var subDir = reporterConfig.subdir || config.subdir
-          var simpleOutputDir = generateOutputDir(browser.name, mainDir, subDir)
-          var resolvedOutputDir = path.resolve(basePath, simpleOutputDir)
-
-          var outputDir = helper.normalizeWinPath(resolvedOutputDir)
-          var options = helper.merge({
-            sourceStore: _.isEmpty(sourceCache) ? null : new SourceCacheStore({
-              sourceCache: sourceCache
-            })
-          }, config, reporterConfig, {
-            dir: outputDir
-          })
-          var reporter = istanbul.Report.create(reporterConfig.type || 'html', options)
-
-          // If reporting to console, skip directory creation
-          if (reporterConfig.type && reporterConfig.type.match(/^(text|text-summary)$/) && typeof reporterConfig.file === 'undefined') {
-            writeReport(reporter, collector)
-            return
-          }
-
-          helper.mkdirIfNotExists(outputDir, function () {
-            log.debug('Writing coverage to %s', outputDir)
-            writeReport(reporter, collector)
-            disposeCollectors()
-          })
+        if (!collector) {
+          return
         }
 
+        pendingFileWritings++
+
+        var mainDir = reporterConfig.dir || config.dir
+        var subDir = reporterConfig.subdir || config.subdir
+        var simpleOutputDir = generateOutputDir(browser.name, mainDir, subDir)
+        var resolvedOutputDir = path.resolve(basePath, simpleOutputDir)
+
+        var outputDir = helper.normalizeWinPath(resolvedOutputDir)
+        var sourceStore = _.isEmpty(sourceCache) ? null : new SourceCacheStore({
+          sourceCache: sourceCache
+        })
+        var options = helper.merge({
+          sourceStore: sourceStore
+        }, config, reporterConfig, {
+          dir: outputDir
+        })
+        var reporter = istanbul.Report.create(reporterConfig.type || 'html', options)
+
+        // If reporting to console, skip directory creation
+        var isConsole = reporterConfig.type && reporterConfig.type.match(/^(text|text-summary)$/)
+        var hasNoFile = _.isUndefined(reporterConfig.file)
+
+        if (isConsole && hasNoFile) {
+          writeReport(reporter, collector)
+          return
+        }
+
+        helper.mkdirIfNotExists(outputDir, function () {
+          log.debug('Writing coverage to %s', outputDir)
+          writeReport(reporter, collector)
+          disposeCollectors()
+        })
       })
     })
